+++ tdbstore/src/main/java/com/deere/ld4mbse/rdfstore/rest/JAXRSApplication.java
@@ -1,6 +1,6 @@
-package com.deere.ld4mbse.vnsm.rest;
+package com.deere.ld4mbse.rdfstore.rest;
 
-import com.deere.ld4mbse.vnsm.model.Environment;
+import com.deere.ld4mbse.rdfstore.model.Environment;
 import javax.ws.rs.ApplicationPath;
 import javax.ws.rs.core.Application;
