+++ tdbstore/src/main/webapp/WEB-INF/web.xml
@@ -11,16 +11,16 @@
     <listener>
         <description>Application verifier and resources releaser</description>
         <display-name>ApplicationListener</display-name>
-        <listener-class>com.deere.ld4mbse.vnsm.web.ApplicationListener</listener-class>
+        <listener-class>com.deere.ld4mbse.rdfstore.web.ApplicationListener</listener-class>
     </listener>
     <resource-env-ref>
-      <description>CDI Bean Manager</description>
-      <resource-env-ref-name>BeanManager</resource-env-ref-name>
-      <resource-env-ref-type>javax.enterprise.inject.spi.BeanManager</resource-env-ref-type>
+        <description>CDI Bean Manager</description>
+        <resource-env-ref-name>BeanManager</resource-env-ref-name>
+        <resource-env-ref-type>javax.enterprise.inject.spi.BeanManager</resource-env-ref-type>
     </resource-env-ref>
     <resource-env-ref>
-      <description>Apache Jena Dataset Singleton Factory</description>
-      <resource-env-ref-name>${tdb.naming.factory}</resource-env-ref-name>
-      <resource-env-ref-type>com.deere.ld4mbse.vnsm.services.DatasetProducer$Factory</resource-env-ref-type>
+        <description>Apache Jena Dataset Singleton Factory</description>
+        <resource-env-ref-name>${tdb.naming.factory}</resource-env-ref-name>
+        <resource-env-ref-type>com.deere.ld4mbse.vnsm.services.DatasetProducer$Factory</resource-env-ref-type>
     </resource-env-ref>
 </web-app>
