+++ tdbstore/src/test/java/com/deere/ld4mbse/rdfstore/services/TDBManagerTest.java
@@ -1,4 +1,4 @@
-package com.deere.ld4mbse.vnsm.services;
+package com.deere.ld4mbse.rdfstore.services;
 
 import org.junit.Before;
